Extract AppHeader and RunButton components for better architecture
- Create AppHeader.vue with title, subtitle, and GitHub link functionality
- Create RunButton.vue with dynamic state and version display
- Remove 445 lines from App.vue (57% reduction: 776→331 lines)
- Clean 2-level component hierarchy with focused responsibilities
- Each component handles its own state and styling

Components now extracted:
- AppHeader: GitHub API, stars, title display
- RunButton: execution state, version info, dynamic styling
- ConsoleOutput: message overlay display
- ConsoleToggle: console toggle button with error states
